requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-socketio>=5.9.0
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
